# chatbot-vf/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación Diputación Alicante",
        "description": "Apis Alicante",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/saveSettingsApi": {
            "post": {
                "tags": [
                    "settings"
                ],
                "summary": "Guarda la configuración de ajustes del cliente.",
                "description": "Guarda la configuración de ajustes del cliente.",
                "operationId": "c05d2681e98bead405375ce799f23599",
                "requestBody": {
                    "description": "Datos de configuración de ajustes del cliente.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city_council_id"
                                ],
                                "properties": {
                                    "city_council_id": {
                                        "description": "ID del cliente",
                                        "type": "integer"
                                    },
                                    "logo": {
                                        "description": "Archivo del logo en formato de imagen",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "color": {
                                        "description": "Color del cliente",
                                        "type": "string"
                                    },
                                    "pie_de_página": {
                                        "description": "Texto del pie de página",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Configuración del cliente guardada exitosamente."
                    },
                    "400": {
                        "description": "Los datos proporcionados son incorrectos o incompletos."
                    },
                    "500": {
                        "description": "Ha ocurrido un error al intentar guardar la configuración del cliente."
                    }
                }
            }
        },
...
        "/api/updateSettingsCityCouncilApi": {
            "post": {
                "tags": [
                    "settings"
                ],
                "summary": "Actualiza la configuración del cliente",
                "description": "Actualiza la configuración del cliente.",
                "operationId": "5262166fe84087ef89425b5988b419bb",
                "requestBody": {
                    "description": "Datos de configuración del cliente",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city_council_id"
                                ],
                                "properties": {
                                    "city_council_id": {
                                        "description": "ID del cliente",
                                        "type": "integer"
                                    },
                                    "logo": {
                                        "description": "Archivo del logo en formato de imagen",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "color": {
                                        "description": "Color del cliente",
                                        "type": "string"
                                    },
                                    "pie_de_página": {
                                        "description": "Texto del pie de página",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Configuración del cliente actualizada exitosamente."
                    },
                    "400": {
                        "description": "Los datos proporcionados son incorrectos o incompletos."
                    },
                    "404": {
                        "description": "La configuración del cliente no fue encontrada."
                    },
                    "500": {
                        "description": "Ha ocurrido un error al intentar actualizar la configuración del cliente."
                    }
                }
            }
        },
        "/api/deleteSettingsApi/{cityCouncilId}": {
            "delete": {
                "tags": [
                    "settings"
                ],
                "summary": "Elimina la configuración de ajustes del cliente.",
                "description": "Elimina la configuración de ajustes del cliente.",
                "operationId": "0f2ba3e501697a2c3b711d0e9fbdce2f",
                "parameters": [
                    {
                        "name": "cityCouncilId",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registros eliminados correctamente."
                    },
                    "404": {
                        "description": "La configuración del cliente no fue encontrada."
                    },
                    "500": {
                        "description": "Ha ocurrido un error al intentar eliminar la configuración del cliente."
                    }
                }
            }
        }
    }
}